ccls: reduce the completion list
Functions or types with default arguments generate one more item per
default argument. This is quickly getting too verbose and a lot of types
from standard c++ library, e.g. vector or map use default type
arguments. The preferred completion is the one with all non-default
arguments which is often not what we want. Disable this feature for now.