zsh: fix fzf history widget
History widget, bound to CTRL-R, duplicates the inserted entry but only
inside tmux and when powerlevel10k theme is used. This commit tries to
fix this issue but changing the function that fzf widgets use
internally. The original function returns either fzf-tmux or fzf but
we've already aliased fzf to fzf-tmux so it should not break anything.