emacs: add selectrum package
Package selectrum is an incremental completion and narrowing framework.
Like Ivy and Helm, which it improves on, Selectrum provides a user
interface for choosing from a list of options by typing a query to
narrow the list, and then selecting one of the remaining candidates.
This offers a significant improvement over the default Emacs interface
for candidate selection.